botv2/requirements.txt
slack-bolt
python-dotenv
PyGithub
Office365-REST-Python-Client